+++ cpp/qtrlib/columns_selecting/ColumnsSelectionUtils.h
@@ -8,6 +8,6 @@ namespace qtr {
 
     std::vector<std::vector<bool>> fingerprintsToColumns(const IndigoFingerprintTable &fingerprints);
 
-    std::vector<int> sortIndexesByValues(const std::vector<double> &values);
+    std::vector<size_t> sortIndexesByValues(const std::vector<double> &values);
 
 } // namespace qtr
